Add jOOQ DSLContext factory
Provides easy access to jOOQ `DSLContext` objects.
